requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlmodel==0.0.21
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0 
python-multipart==0.0.6 # Add this line for JWT support